+++ src/model/trade.js
@@ -66,7 +66,7 @@ export default class Trade {
 	}
 	
 	isProfitable(){
-		return this.profit >= 0;
+		return this.profit > 0;
 	}
 	
 	calculateProfit(){