+++ src/service/satoshiAPI.js
@@ -21,12 +21,13 @@ let TradeSatoshi = () => {
 			    'Content-Type': 'application/json; charset=utf-8'
 		    },
 		    body: JSON.stringify(params),
-        maxAttempts: 10,
+        maxAttempts: 20,
         retryDelay: 3000
 	    };
 	    //Need to put in retries here.....
         try {
             let response = await request.post(reqOpts);
+            //console.log(response);
 	          response = JSON.parse(response.body);
             return response.success ? Promise.resolve(response.result) : Promise.reject(response.message);
           
@@ -108,7 +109,7 @@ let TradeSatoshi = () => {
         submitOrder: async (params = {}) => {
             if (!params.market) {
                 return Promise.reject("submitOrder(), You must supply a valid market or trade pair Id!");
-            } else if (params.type && params.type !== 'Buy' && params.type !== 'Sel') {
+            } else if (params.type && params.type !== 'Buy' && params.type !== 'Sell') {
                 return Promise.reject("submitOrder(), You must supply a valid Type, e.g. 'Sell' or 'Buy'!");
             }  else if (!params.amount || typeof params.amount !== 'number') {
 	              return Promise.reject("submitOrder(), You must supply a valid Amount, e.g. Amount: '123.00000000'!");